refactor(horizontalscrollindicator): rewrite to agreed format
