+++ source/components/atoms/HorizontalScrollIndicator/HorizontalScrollIndicator.styled.types.ts
@@ -1,0 +1,4 @@
+export interface Props {
+  percentage: number;
+  numberOfSegments?: number;
+}
