+++ source/components/atoms/HorizontalScrollIndicator/HorizontalScrollIndicator.tsx
@@ -1,5 +1,13 @@
 import React from "react";
-import styled from "styled-components/native";
+
+import {
+  SegmentWrapper,
+  StartSegment,
+  EndSegment,
+  MiddleSegment,
+} from "./HorizontalScrollIndicator.styled";
+
+import type { Props } from "./HorizontalScrollIndicator.styled.types";
 
 function* range(start: number, end: number) {
   for (let i = start; i <= end; i++) {
@@ -7,41 +15,7 @@ function* range(start: number, end: number) {
   }
 }
 
-const SegmentWrapper = styled.View`
-  flex-direction: row;
-  justify-content: center;
-`;
-
-const StartSegment = styled.View<{ selected?: boolean }>`
-  height: 3px;
-  width: 12px;
-  margin-right: 3px;
-  border-top-left-radius: 2px;
-  border-bottom-left-radius: 2px;
-  background-color: ${({ selected, theme }) =>
-    selected ? theme.colors.neutrals[2] : theme.colors.neutrals[4]};
-`;
-const EndSegment = styled.View<{ selected?: boolean }>`
-  height: 3px;
-  width: 12px;
-  border-top-right-radius: 2px;
-  border-bottom-right-radius: 2px;
-  background-color: ${({ selected, theme }) =>
-    selected ? theme.colors.neutrals[2] : theme.colors.neutrals[4]};
-`;
-const MiddleSegment = styled.View<{ selected?: boolean }>`
-  height: 3px;
-  width: 12px;
-  margin-right: 3px;
-  background-color: ${({ selected, theme }) =>
-    selected ? theme.colors.neutrals[2] : theme.colors.neutrals[4]};
-`;
-interface CustomScrollIndicatorProps {
-  percentage: number;
-  numberOfSegments?: number;
-}
-
-const HorizontalScrollIndicator: React.FC<CustomScrollIndicatorProps> = ({
+const HorizontalScrollIndicator: React.FC<Props> = ({
   percentage,
   numberOfSegments = 3,
 }) => {
